grafana/consumer_app/requirements.txt
kafka-python
prometheus-client
